Rename endian.h to endian_util.h to avoid conflicts with Linux headers
